ml-api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
transformers==4.40.2
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
python-dotenv==1.0.1
numpy==1.26.4
